+++ image/build.sh
@@ -1,15 +1,9 @@
 #!/bin/bash -ex
 
 ## Add bash tools to /sbin
-ln -s /container/tool/add-host /sbin/add-host
-ln -s /container/tool/install-multiple-process-stack /sbin/install-multiple-process-stack
-ln -s /container/tool/install-service /sbin/install-service
-ln -s /container/tool/install-service-available /sbin/install-service-available
-ln -s /container/tool/remove-service /sbin/remove-service
-ln -s /container/tool/run /sbin/run
+ln -s /container/tool/* /sbin/
 
-# Add python tools and needed directories
-ln -s /container/tool/py_tool/my_init /sbin/my_init
+# Add needed directories & files
 mkdir -p /etc/service
 mkdir -p /etc/my_init.d
 mkdir -p /etc/container_environment
@@ -20,15 +14,9 @@ groupadd -g 8377 docker_env
 chown :docker_env /etc/container_environment.sh
 chmod 640 /etc/container_environment.sh
 
-ln -s /container/tool/py_tool/setuser /sbin/setuser
-
-# dpkg
+# dpkg options
 cp /container/file/dpkg_nodoc /etc/dpkg/dpkg.cfg.d/01_nodoc
 cp /container/file/dpkg_nolocales /etc/dpkg/dpkg.cfg.d/01_nolocales
-
-# Remove useless files
-rm -rf /container/file
-rm -rf /container/build.sh /container/Dockerfile
 
 # General config
 export LC_ALL=C
@@ -62,12 +50,32 @@ dpkg-divert --local --rename --add /usr/bin/ischroot
 ln -sf /bin/true /usr/bin/ischroot
 
 ## Install apt-utils.
-$minimal_apt_get_install apt-utils
+$minimal_apt_get_install apt-utils python locales
 
 ## Upgrade all packages.
 apt-get dist-upgrade -y --no-install-recommends
+
+# fix locale
+locale-gen en_US.UTF-8 en_us
+locale-gen C.UTF-8
+dpkg-reconfigure locales
+/usr/sbin/update-locale LANG=C.UTF-8
+
+echo -n C.UTF-8 > /etc/container_environment/LANG
+echo -n C.UTF-8 > /etc/container_environment/LANGUAGE
+echo -n C.UTF-8 > /etc/container_environment/LC_CTYPE
+
+# install PyYAML
+tar -C /container/file/ -xvf /container/file/PyYAML-3.11.tar.gz
+cd /container/file/PyYAML-3.11/
+python setup.py install
+cd -
 
 apt-get clean
 rm -rf /tmp/* /var/tmp/*
 rm -rf /var/lib/apt/lists/*
 rm -f /etc/dpkg/dpkg.cfg.d/02apt-speedup
+
+# Remove useless files
+rm -rf /container/file
+rm -rf /container/build.sh /container/Dockerfile