+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # Changelog
 
+## 0.1.1
+  - Fix remove-service #1
+  - Add python and PyYAML
+  - Fix locales
+  - Fix my_init
+
 ## 0.1.0
   - Initial release